refactor(gui): 重命名页面布局样式模块
Why:
- `style.py` 作为页面级布局与壳层 helper，名称过于宽泛，和 `styles.py` 容易混淆
- 将其改成 `page_style.py` 后，更符合当前模块职责与后续阅读预期

What:
- 将 `gui/common/style.py` 重命名为 `gui/common/page_style.py`
- 同步 about、execution、home、overview、setting 五个页面的导入路径

Risk:
- 兼容性影响：无，仅调整模块名与导入路径，不改运行时行为
- 潜在风险：若仍有遗漏导入，页面初始化时会暴露 ImportError

Test:
- 运行 uv run pytest -q tests/gui/test_about_page.py tests/gui/test_execution_page.py tests/gui/test_home_page.py tests/gui/test_overview_panels.py tests/gui/test_setting_page_theme_preset.py tests/gui/test_setting_panels_module.py，39 passed in 27.40s
