+++ src/lol_audio_unpack/gui/view/execution_page.py
@@ -18,7 +18,7 @@ from lol_audio_unpack.gui.common import (
     get_buffered_log_lines,
     show_feedback_infobar,
 )
-from lol_audio_unpack.gui.common.style import (
+from lol_audio_unpack.gui.common.page_style import (
     apply_page_content_margins,
     configure_transparent_scroll_page,
 )
