+++ src/lol_audio_unpack/gui/view/home_page.py
@@ -26,7 +26,7 @@ from qfluentwidgets import (
 from qfluentwidgets import FluentIcon as FIF
 
 from lol_audio_unpack.gui.common import GuiConfig, format_default_relative_path
-from lol_audio_unpack.gui.common.style import (
+from lol_audio_unpack.gui.common.page_style import (
     apply_page_content_margins,
     configure_transparent_scroll_page,
 )
